r851: Merge 2.1: Here is another load of small changes. * The cursor handling was changed (it is hidden and shown at some places). * Added some (currently dummy) arguments to functions in maincursor.C. * AudioALSA::list_devices() now has an argument for selecting between   recording and playback devices. * Rounding error in picon drawing for long timelines fixed. * Additional format presets. * Assorted whitespace and code moving changes. * An ID for resources is introduced. * Implement render_stop() that cleans up after plugins (for live plugins). * Moved video device creation from open_input() to new_device_base(). * VideoDevice::open_input() takes the frame rate as double instead of float. * PictureConfig does not need an MWindow pointer, a BC_Hash* is enough. * Remove video_is_on() and replace it by get_video_on(). and probably a lot more.
This revision should match GIT::3c7a200ee98947c540d802853ead4602d4f16574.

+++ cinelerra/filebase.h
@@ -60,7 +60,7 @@ public:
 
 // Subclass should call this to add the base class allocation.
 // Only used in read mode.
-	virtual int get_memory_usage() { return 0; };
+	virtual int64_t get_memory_usage() { return 0; };
 
 	virtual int write_samples(double **buffer, 
 		int64_t len) { return 0; };
