+++ cinelerra/filetga.C
@@ -169,9 +169,9 @@ FrameWriterUnit* FileTGA::new_writer_unit(FrameWriter *writer)
 	return new TGAUnit(this, writer);
 }
 
-int FileTGA::get_memory_usage()
+int64_t FileTGA::get_memory_usage()
 {
-	int result = FileList::get_memory_usage();
+	int64_t result = FileList::get_memory_usage();
 	if(temp) result += temp->get_data_size();
 	return result;
 }
